organizing and restructuring the backend folder
